+reset(); rewrite restore() for settings data deletions
The reset() method will remove all or some of the stored keys.  This is useful for testing.
The restore() method was not robust in all the cases where Settings was used to modify the stored data.  Clearing cached data, clearing data, and deleting the app left some inconsistent state.  The fix was to totally restructure this so that a failed get could restore the default.  I also broke out the put for default values so all state can go into a single object and be set just once.